Python Sample_CA Liquor analysis
Greetings and welcome to this notebook, designed to demonstrate the application of advanced analytics in the liquor market. The dataset under consideration, though simulated, is crafted to mirror the characteristics of the California market, thereby furnishing realistic insights into the portfolio of Brown-Forman and the wider market landscape.

Our journey begins with a comprehensive exploration of Brown-Forman's portfolio performance, expanding thereafter to a holistic market overview. In the subsequent section, we shall delve deep into the pricing strategies of key products, crystallizing our analysis with sales forecasts for the next 12 months using state-of-the-art time series models.

For the sake of brevity and to maintain the focus on central themes, certain analytical procedures such as data collection, cleansing, hypothesis testing, and model evaluation and tuning have been abbreviated or intentionally excluded. However, these steps are crucial and are invariably incorporated in a full-scale analysis.
Please note that the data utilized here is synthetic, meticulously crafted to replicate real-world scenarios. We believe this approach provides a practical, yet simplified, exploration of analytical methodologies in action. 

Thank you for your time and interest.